backend/requirements.txt
# Flask Framework
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0

# File Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-magic==0.4.27
pdfminer.six==20221105

# NLP & AI
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
nltk==3.8.1
openai==1.3.7

# Data Validation & Processing
marshmallow==3.20.1
python-dateutil==2.8.2

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
Werkzeug==3.0.1
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
